requirements.txt
requests==2.25.1
beautifulsoup4==4.9.3
mysql-connector-python==8.0.23
libsql-experimental
python-dotenv
pdfminer.six
mysql-connector-python